+++ internal/pkg/service/models/input.go
@@ -7,22 +7,22 @@ type AddTaskData struct {
 
 // UpdateTaskData структура запроса на обновление задачи
 type UpdateTaskData struct {
-	ID                 uint
+	ID                 uint64
 	Title, Description string
 }
 
 // ListTaskData структура запроса на получение списка задач
 type ListTaskData struct {
-	MaxTasksCount uint
-	Offset        uint
+	MaxTasksCount uint64
+	Offset        uint64
 }
 
 // DeleteTaskData структура запроса на удаления задачи
 type DeleteTaskData struct {
-	ID uint
+	ID uint64
 }
 
 // GetTaskData структура запроса на получение описания задачи
 type GetTaskData struct {
-	ID uint
+	ID uint64
 }
